refactor(storage): split the legacy schema rebuilds out of shared_db
ARCH-001: shared_db.rs had grown to 1336 non-test lines against a 1000-line
cap. Move the canonical mail_messages DDL macros, the message_text nullability
rebuild and its eight unit tests into `mail_messages_schema`, the team_roster
harness rebuild and its tests into `team_roster_schema`, and the pieces both
rebuilds share into `schema_support`. shared_db.rs keeps DB_MIGRATIONS (which
still expands the shared DDL macros) and the `ensure_schema` call sites, and
is now 919 non-test lines. Behaviour is unchanged.

RBQA-F002: the team_roster rebuild copies rows by explicit column list and so
could silently drop an unknown legacy column, the same defect the mail_messages
rebuild guards against. Both now call
`schema_support::ensure_columns_match_canonical`, which rejects unknown and
missing columns by name before any write.

RBQA-F001: when a rebuild failed and restoring `PRAGMA foreign_keys = ON` also
failed, the restore failure was dropped. `schema_support::merge_restore_failure`
now folds it into the original error's preserved cause, keeping the original
code and message.
