+++ crates/atm-storage-rusqlite/src/shared_db.rs
@@ -1,3 +1,4 @@
+use crate::mail_messages_schema::{mail_messages_index_ddl, mail_messages_table_ddl};
 pub(crate) use crate::shared_db_reader_lanes::SharedDb;
 use crate::writer::{WriteOp, WriteOpResult, validate_upsert_message_request};
 use atm_storage::AsyncMailboxReader;
@@ -16,7 +17,6 @@ use std::sync::Arc;
 use std::sync::atomic::{AtomicU64, Ordering};
 #[cfg(test)]
 use std::sync::{LazyLock, Mutex};
-use std::time::Instant;
 
 // Search projection writes touch several B-trees and FTS segments inside the
 // sole durable writer transaction.  The SQLite default is only 2 MiB, which
@@ -24,97 +24,10 @@ use std::time::Instant;
 // only to the one writer connection and bounds its page cache at 32 MiB.
 const WRITER_CACHE_KIB: i64 = 32 * 1024;
 #[cfg(test)]
-static NEXT_IN_MEMORY_DB_ID: AtomicU64 = AtomicU64::new(1);
+pub(crate) static NEXT_IN_MEMORY_DB_ID: AtomicU64 = AtomicU64::new(1);
 #[cfg(test)]
 static OPENED_CONNECTIONS: LazyLock<Mutex<std::collections::HashMap<String, usize>>> =
     LazyLock::new(|| Mutex::new(std::collections::HashMap::new()));
-
-/// Canonical `mail_messages` table DDL, parameterized by its `CREATE` clause.
-///
-/// The fresh-database schema in [`DB_MIGRATIONS`] and the legacy
-/// `message_text` rebuild both materialize their table from this single
-/// definition, so the two shapes cannot drift apart. The column order here is
-/// also the order a fresh database ends up with after the additive
-/// `ALTER TABLE` migrations in `ensure_mail_message_columns` and
-/// `template_catalog_schema::ensure_schema` have run.
-macro_rules! mail_messages_table_ddl {
-    ($create:literal) => {
-        concat!(
-            $create,
-            r#" (
-    team TEXT NOT NULL,
-    agent TEXT NOT NULL,
-    message_key TEXT NOT NULL,
-    envelope_json TEXT NOT NULL,
-    from_agent TEXT NOT NULL,
-    source_chat_id TEXT NULL,
-    destination_chat_id TEXT NULL,
-    message_text TEXT NULL,
-    template_sha TEXT NULL,
-    vars_json TEXT NULL,
-    category TEXT NULL,
-    content_format TEXT NULL,
-    tags_json TEXT NOT NULL DEFAULT '[]',
-    summary TEXT NULL,
-    message_at TEXT NOT NULL,
-    message_id TEXT NULL,
-    parent_message_id TEXT NULL,
-    thread_mode TEXT NULL CHECK(thread_mode IS NULL OR thread_mode IN ('add-details', 'supersede')),
-    recorded_at TEXT,
-    workflow_scope_kind TEXT NULL,
-    workflow_scope_id TEXT NULL,
-    workflow_state TEXT NULL,
-    workflow_stage TEXT NULL,
-    workflow_transition TEXT NULL,
-    workflow_iteration TEXT NULL,
-    applied_template_tags_json TEXT NULL,
-    effective_tags_json TEXT NULL,
-    CHECK(message_key GLOB 'atm:*' OR message_key GLOB 'ext:*'),
-    PRIMARY KEY (team, agent, message_key)
-);
-"#
-        )
-    };
-}
-
-/// Every index owned by `mail_messages`.
-///
-/// The rebuild drops the table (and therefore its indexes), so it replays this
-/// exact definition to guarantee a migrated database is index-identical to a
-/// fresh one.
-macro_rules! mail_messages_index_ddl {
-    () => {
-        r#"
-CREATE UNIQUE INDEX IF NOT EXISTS uq_mail_messages_single_successor
-    ON mail_messages(team, agent, parent_message_id)
-    WHERE parent_message_id IS NOT NULL;
-
-CREATE UNIQUE INDEX IF NOT EXISTS uq_mail_messages_message_id
-    ON mail_messages(team, agent, message_id)
-    WHERE message_id IS NOT NULL;
-
-CREATE INDEX IF NOT EXISTS idx_mail_messages_mailbox
-    ON mail_messages(team, agent);
-
--- Post-commit received-hook dispatch reloads an admitted record by its
--- immutable message key.  The mailbox primary key begins with team and agent,
--- so it cannot serve that global-key lookup without scanning a growing table.
-CREATE INDEX IF NOT EXISTS idx_mail_messages_message_key
-    ON mail_messages(message_key);
-"#
-    };
-}
-
-/// Staging table name used by the `message_text` nullability rebuild.
-const MAIL_MESSAGES_REBUILD_TABLE: &str = "mail_messages_nullable_rebuild";
-
-/// `CREATE TABLE` for the rebuild staging table, byte-identical in shape to
-/// the fresh `mail_messages` definition.
-const MAIL_MESSAGES_REBUILD_TABLE_DDL: &str =
-    mail_messages_table_ddl!("CREATE TABLE mail_messages_nullable_rebuild");
-
-/// Index DDL replayed after the rebuild swaps the tables.
-const MAIL_MESSAGES_INDEX_DDL: &str = mail_messages_index_ddl!();
 
 pub(crate) const DB_MIGRATIONS: &str = concat!(
     mail_messages_table_ddl!("CREATE TABLE IF NOT EXISTS mail_messages"),
@@ -658,10 +571,10 @@ pub(crate) fn ensure_schema(
     // Runs after every additive column migration so the rebuilt table can be
     // populated from a legacy table that already carries the current column
     // set, and before the search projections, which read `mail_messages`.
-    ensure_mail_messages_message_text_nullable(connection, target)?;
+    crate::mail_messages_schema::ensure_mail_messages_message_text_nullable(connection, target)?;
     crate::search_schema::ensure_schema(connection, target)?;
     ensure_team_roster_columns(connection, target)?;
-    ensure_team_roster_harness_values(connection, target)?;
+    crate::team_roster_schema::ensure_team_roster_harness_values(connection, target)?;
     ensure_team_nudge_template_override_columns(connection, target)?;
     ensure_mail_message_states_nudge_columns(connection, target)?;
     crate::graft_receiver_endpoint_schema::ensure_schema(connection, target)?;
@@ -730,262 +643,6 @@ fn ensure_mail_message_columns(
     )
 }
 
-/// Relaxes a legacy `mail_messages.message_text NOT NULL` column to nullable.
-///
-/// Databases created before decomposed template admission shipped still carry
-/// `message_text TEXT NOT NULL`. SQLite cannot relax a `NOT NULL` constraint
-/// in place, and the additive `ensure_column` migrations only add missing
-/// columns, so `atm send --template` fails on those databases with a
-/// constraint violation when the decomposed writer nulls the column out.
-///
-/// The probe is self-detecting rather than version-stamped: it reads the live
-/// column flag and returns without writing anything when the table is missing
-/// or the column is already nullable, so correct databases pay one pragma
-/// query per startup.
-///
-/// # Errors
-/// Returns an error when the probe or any rebuild stage fails. Every failure
-/// rolls the rebuild transaction back, so the legacy table is left untouched
-/// and a later startup retries.
-fn ensure_mail_messages_message_text_nullable(
-    connection: &mut Connection,
-    target: &SharedDbTarget,
-) -> Result<(), AtmError> {
-    use rusqlite::OptionalExtension;
-
-    let legacy_not_null = connection
-        .query_row(
-            r#"SELECT "notnull" FROM pragma_table_info('mail_messages') WHERE name = 'message_text';"#,
-            [],
-            |row| row.get::<_, i64>(0),
-        )
-        .optional()
-        .map_err(|error| {
-            sqlite_error(
-                target,
-                "failed to probe mail_messages.message_text nullability",
-                error,
-            )
-        })?;
-    if legacy_not_null != Some(1) {
-        return Ok(());
-    }
-
-    // `PRAGMA foreign_keys` is a no-op inside a transaction, so it must be
-    // toggled here: `mail_message_states` has a foreign key onto
-    // `mail_messages`, and the drop/rename below would otherwise cascade the
-    // state rows away. This follows SQLite's documented ALTER TABLE procedure.
-    connection
-        .pragma_update(None, "foreign_keys", "OFF")
-        .map_err(|error| {
-            sqlite_error(
-                target,
-                "failed to disable foreign keys for the mail_messages rebuild",
-                error,
-            )
-        })?;
-    let rebuild = rebuild_mail_messages_with_nullable_message_text(connection, target);
-    let restored = connection
-        .pragma_update(None, "foreign_keys", "ON")
-        .map_err(|error| {
-            sqlite_error(
-                target,
-                "failed to restore foreign keys after the mail_messages rebuild",
-                error,
-            )
-        });
-    rebuild?;
-    restored
-}
-
-/// Performs the SQLite table rebuild that relaxes `message_text` to nullable.
-///
-/// Callers must have disabled `PRAGMA foreign_keys` and must restore it
-/// afterwards; see [`ensure_mail_messages_message_text_nullable`].
-fn rebuild_mail_messages_with_nullable_message_text(
-    connection: &mut Connection,
-    target: &SharedDbTarget,
-) -> Result<(), AtmError> {
-    let started = Instant::now();
-    let transaction = connection
-        .transaction_with_behavior(TransactionBehavior::Immediate)
-        .map_err(|error| {
-            sqlite_error(
-                target,
-                "failed to start the mail_messages message_text rebuild",
-                error,
-            )
-        })?;
-
-    let legacy_columns = table_column_names(&transaction, target, "mail_messages")?;
-
-    // The catalog view selects from `mail_messages`; SQLite validates view
-    // bodies during `ALTER TABLE ... RENAME`, so it is dropped up front and
-    // recreated verbatim from its owning module once the swap is done.
-    transaction
-        .execute_batch("DROP VIEW IF EXISTS decomposed_messages;")
-        .map_err(|error| {
-            sqlite_error(
-                target,
-                "failed to drop the decomposed_messages view for the mail_messages rebuild",
-                error,
-            )
-        })?;
-
-    transaction
-        .execute_batch(MAIL_MESSAGES_REBUILD_TABLE_DDL)
-        .map_err(|error| {
-            sqlite_error(
-                target,
-                "failed to create the rebuilt mail_messages table",
-                error,
-            )
-        })?;
-    let rebuilt_columns = table_column_names(&transaction, target, MAIL_MESSAGES_REBUILD_TABLE)?;
-
-    let unknown = difference(&legacy_columns, &rebuilt_columns);
-    if !unknown.is_empty() {
-        return Err(AtmError::mailbox_write(format!(
-            "refusing to rebuild mail_messages on {}: the legacy table carries column(s) {} that the current schema does not define",
-            target.display(),
-            unknown.join(", ")
-        )));
-    }
-    let absent = difference(&rebuilt_columns, &legacy_columns);
-    if !absent.is_empty() {
-        return Err(AtmError::mailbox_write(format!(
-            "refusing to rebuild mail_messages on {}: the legacy table is missing column(s) {} that the additive migrations should have added",
-            target.display(),
-            absent.join(", ")
-        )));
-    }
-
-    let column_list = rebuilt_columns
-        .iter()
-        .map(|column| format!("\"{column}\""))
-        .collect::<Vec<_>>()
-        .join(", ");
-    let copied = transaction
-        .execute(
-            &format!(
-                "INSERT INTO {MAIL_MESSAGES_REBUILD_TABLE} ({column_list}) SELECT {column_list} FROM mail_messages;"
-            ),
-            [],
-        )
-        .map_err(|error| {
-            sqlite_error(
-                target,
-                "failed to copy rows into the rebuilt mail_messages table",
-                error,
-            )
-        })?;
-
-    transaction
-        .execute_batch(&format!(
-            "DROP TABLE mail_messages;
-             ALTER TABLE {MAIL_MESSAGES_REBUILD_TABLE} RENAME TO mail_messages;"
-        ))
-        .map_err(|error| {
-            sqlite_error(
-                target,
-                "failed to swap in the rebuilt mail_messages table",
-                error,
-            )
-        })?;
-
-    transaction
-        .execute_batch(MAIL_MESSAGES_INDEX_DDL)
-        .map_err(|error| {
-            sqlite_error(
-                target,
-                "failed to recreate mail_messages indexes after the rebuild",
-                error,
-            )
-        })?;
-    crate::template_catalog_schema::ensure_schema(&transaction, target)?;
-
-    let violations = transaction
-        .query_row(
-            "SELECT COUNT(1) FROM pragma_foreign_key_check;",
-            [],
-            |row| row.get::<_, i64>(0),
-        )
-        .map_err(|error| {
-            sqlite_error(
-                target,
-                "failed to run the foreign-key check after the mail_messages rebuild",
-                error,
-            )
-        })?;
-    if violations > 0 {
-        return Err(AtmError::mailbox_write(format!(
-            "mail_messages rebuild on {} left {violations} foreign key violation(s); rolling back",
-            target.display()
-        )));
-    }
-
-    transaction.commit().map_err(|error| {
-        sqlite_error(
-            target,
-            "failed to commit the mail_messages message_text rebuild",
-            error,
-        )
-    })?;
-    tracing::info!(
-        event = "sqlite.mail_messages.message_text_rebuild",
-        db.namespace = %target.display(),
-        db.rows_copied = copied,
-        db.elapsed_ms = u64::try_from(started.elapsed().as_millis()).unwrap_or(u64::MAX),
-        "rebuilt mail_messages so message_text is nullable",
-    );
-    Ok(())
-}
-
-/// Returns the entries of `left` that do not appear in `right`.
-fn difference(left: &[String], right: &[String]) -> Vec<String> {
-    left.iter()
-        .filter(|column| !right.contains(column))
-        .cloned()
-        .collect()
-}
-
-/// Reads the declared column names of `table` in their declaration order.
-fn table_column_names(
-    connection: &Connection,
-    target: &SharedDbTarget,
-    table: &str,
-) -> Result<Vec<String>, AtmError> {
-    let mut statement = connection
-        .prepare(&format!("PRAGMA table_info({table});"))
-        .map_err(|error| {
-            sqlite_error(
-                target,
-                format!("failed to inspect sqlite table {table}"),
-                error,
-            )
-        })?;
-    let columns = statement
-        .query_map([], |row| row.get::<_, String>(1))
-        .map_err(|error| {
-            sqlite_error(
-                target,
-                format!("failed to enumerate sqlite columns for {table}"),
-                error,
-            )
-        })?;
-    columns
-        .map(|entry| {
-            entry.map_err(|error| {
-                sqlite_error(
-                    target,
-                    format!("failed to read sqlite column metadata for {table}"),
-                    error,
-                )
-            })
-        })
-        .collect()
-}
-
 fn ensure_team_roster_columns(
     connection: &Connection,
     target: &SharedDbTarget,
@@ -1025,80 +682,6 @@ fn ensure_team_roster_columns(
         "metadata_json",
         "ALTER TABLE team_roster ADD COLUMN metadata_json TEXT NOT NULL DEFAULT '{}';",
     )
-}
-
-fn ensure_team_roster_harness_values(
-    connection: &mut Connection,
-    target: &SharedDbTarget,
-) -> Result<(), AtmError> {
-    let table_sql = connection
-        .query_row(
-            "SELECT sql FROM sqlite_master WHERE type = 'table' AND name = 'team_roster';",
-            [],
-            |row| row.get::<_, String>(0),
-        )
-        .map_err(|error| {
-            sqlite_error(
-                target,
-                "failed to inspect team_roster harness constraint",
-                error,
-            )
-        })?;
-    if table_sql.contains("'hermes'") && table_sql.contains("'python-graft'") {
-        return Ok(());
-    }
-
-    let transaction = connection
-        .transaction_with_behavior(TransactionBehavior::Immediate)
-        .map_err(|error| {
-            sqlite_error(
-                target,
-                "failed to start team_roster harness migration",
-                error,
-            )
-        })?;
-    transaction
-        .execute_batch(
-            "DROP INDEX IF EXISTS idx_team_roster_team_name;
-             ALTER TABLE team_roster RENAME TO team_roster_legacy_harness;
-             CREATE TABLE team_roster (
-                 team_name TEXT NOT NULL,
-                 agent_name TEXT NOT NULL,
-                 member_kind TEXT NOT NULL CHECK(member_kind IN ('permanent', 'ephemeral')),
-                 harness TEXT NOT NULL CHECK(harness IN ('claude-code', 'codex-cli', 'gemini-cli', 'opencode', 'hermes', 'python-graft')),
-                 agent_type TEXT NOT NULL DEFAULT '',
-                 model TEXT NOT NULL DEFAULT '',
-                 metadata_json TEXT NOT NULL DEFAULT '{}',
-                 source TEXT,
-                 recipient_pane_id TEXT NULL,
-                 updated_at TEXT NOT NULL,
-                 PRIMARY KEY (team_name, agent_name)
-             );
-             INSERT INTO team_roster(
-                 team_name, agent_name, member_kind, harness, agent_type, model,
-                 metadata_json, source, recipient_pane_id, updated_at
-             )
-             SELECT
-                 team_name, agent_name, member_kind, harness, agent_type, model,
-                 metadata_json, source, recipient_pane_id, updated_at
-             FROM team_roster_legacy_harness;
-             DROP TABLE team_roster_legacy_harness;
-             CREATE INDEX idx_team_roster_team_name ON team_roster(team_name);",
-        )
-        .map_err(|error| {
-            sqlite_error(
-                target,
-                "failed to migrate team_roster harness constraint",
-                error,
-            )
-        })?;
-    transaction.commit().map_err(|error| {
-        sqlite_error(
-            target,
-            "failed to commit team_roster harness migration",
-            error,
-        )
-    })
 }
 
 fn ensure_team_nudge_template_override_columns(
@@ -1622,59 +1205,6 @@ mod tests {
     }
 
     #[test]
-    fn ensure_team_roster_harness_values_migrates_legacy_check_constraint() {
-        let target = SharedDbTarget::InMemory {
-            uri: format!(
-                "file:atm-storage-rusqlite-roster-harness-test-{}?mode=memory&cache=shared",
-                NEXT_IN_MEMORY_DB_ID.fetch_add(1, Ordering::Relaxed)
-            ),
-        };
-        let mut connection = open_connection_for_target(&target).expect("open connection");
-        connection
-            .execute_batch(
-                "CREATE TABLE team_roster (
-                    team_name TEXT NOT NULL,
-                    agent_name TEXT NOT NULL,
-                    member_kind TEXT NOT NULL CHECK(member_kind IN ('permanent', 'ephemeral')),
-                    harness TEXT NOT NULL CHECK(harness IN ('claude-code', 'codex-cli', 'gemini-cli', 'opencode')),
-                    agent_type TEXT NOT NULL DEFAULT '',
-                    model TEXT NOT NULL DEFAULT '',
-                    metadata_json TEXT NOT NULL DEFAULT '{}',
-                    source TEXT,
-                    recipient_pane_id TEXT NULL,
-                    updated_at TEXT NOT NULL,
-                    PRIMARY KEY (team_name, agent_name)
-                );
-                CREATE INDEX idx_team_roster_team_name ON team_roster(team_name);
-                INSERT INTO team_roster(
-                    team_name, agent_name, member_kind, harness, updated_at
-                ) VALUES ('hermes', 'skillrx', 'permanent', 'claude-code', 'now');",
-            )
-            .expect("create legacy roster table");
-
-        ensure_team_roster_harness_values(&mut connection, &target).expect("migrate roster");
-        connection
-            .execute(
-                "INSERT INTO team_roster(
-                    team_name, agent_name, member_kind, harness, updated_at
-                ) VALUES ('hermes', 'python-worker', 'permanent', 'python-graft', 'now');",
-                [],
-            )
-            .expect("new harness should satisfy migrated check constraint");
-        let harnesses: Vec<String> = connection
-            .prepare(
-                "SELECT harness FROM team_roster
-                 WHERE team_name = 'hermes' ORDER BY agent_name;",
-            )
-            .expect("prepare harness query")
-            .query_map([], |row| row.get(0))
-            .expect("query harnesses")
-            .collect::<Result<_, _>>()
-            .expect("decode harnesses");
-        assert_eq!(harnesses, vec!["python-graft", "claude-code"]);
-    }
-
-    #[test]
     fn ensure_team_nudge_template_override_columns_migrates_legacy_empty_rows_to_disabled() {
         let target = SharedDbTarget::InMemory {
             uri: format!(
@@ -1822,424 +1352,5 @@ mod tests {
         // database must not error.
         ensure_mail_message_states_nudge_columns(&connection, &target)
             .expect("re-run migration idempotently");
-    }
-
-    /// `mail_messages` exactly as databases created before the nullable
-    /// `message_text` change carry it, including the foreign-key dependent
-    /// state table and every index owned by the mailbox table.
-    const LEGACY_MAIL_SCHEMA_DDL: &str = r#"
-CREATE TABLE mail_messages (
-    team TEXT NOT NULL, agent TEXT NOT NULL, message_key TEXT NOT NULL,
-    envelope_json TEXT NOT NULL, from_agent TEXT NOT NULL,
-    source_chat_id TEXT NULL, destination_chat_id TEXT NULL,
-    message_text TEXT NOT NULL, summary TEXT NULL, message_at TEXT NOT NULL,
-    message_id TEXT NULL, parent_message_id TEXT NULL,
-    thread_mode TEXT NULL CHECK(thread_mode IS NULL OR thread_mode IN ('add-details', 'supersede')),
-    recorded_at TEXT,
-    CHECK(message_key GLOB 'atm:*' OR message_key GLOB 'ext:*'),
-    PRIMARY KEY (team, agent, message_key)
-);
-
-CREATE TABLE mail_message_states (
-    team TEXT NOT NULL,
-    agent TEXT NOT NULL,
-    message_key TEXT NOT NULL,
-    read INTEGER NOT NULL DEFAULT 0 CHECK(read IN (0, 1)),
-    pending_ack_at TEXT NULL,
-    acknowledged_at TEXT NULL,
-    expires_at TEXT NULL,
-    deleted_at TEXT NULL,
-    updated_at TEXT NULL,
-    PRIMARY KEY (team, agent, message_key),
-    FOREIGN KEY (team, agent, message_key)
-        REFERENCES mail_messages(team, agent, message_key)
-        ON DELETE CASCADE
-);
-
-CREATE UNIQUE INDEX uq_mail_messages_single_successor
-    ON mail_messages(team, agent, parent_message_id)
-    WHERE parent_message_id IS NOT NULL;
-
-CREATE UNIQUE INDEX uq_mail_messages_message_id
-    ON mail_messages(team, agent, message_id)
-    WHERE message_id IS NOT NULL;
-
-CREATE INDEX idx_mail_messages_mailbox
-    ON mail_messages(team, agent);
-
-CREATE INDEX idx_mail_messages_message_key
-    ON mail_messages(message_key);
-"#;
-
-    const LEGACY_MAIL_ROWS_DDL: &str = r#"
-INSERT INTO mail_messages(
-    team, agent, message_key, envelope_json, from_agent, source_chat_id,
-    destination_chat_id, message_text, summary, message_at, message_id,
-    parent_message_id, thread_mode, recorded_at
-) VALUES
- ('team-a', 'agent-a', 'atm:one', '{"one":1}', 'sender-one', 'src-1', 'dst-1',
-  'body one', 'summary one', '2026-07-13T00:00:00Z', 'msg-1', NULL, NULL,
-  '2026-07-13T00:00:01Z'),
- ('team-a', 'agent-a', 'atm:two', '{"two":2}', 'sender-two', NULL, NULL,
-  'body two', NULL, '2026-07-13T00:00:02Z', 'msg-2', 'msg-1', 'supersede',
-  '2026-07-13T00:00:03Z'),
- ('team-b', 'agent-b', 'ext:three', '{"three":3}', 'sender-three', NULL, NULL,
-  'body three', 'summary three', '2026-07-13T00:00:04Z', NULL, NULL, NULL, NULL);
-
-INSERT INTO mail_message_states(
-    team, agent, message_key, read, acknowledged_at, updated_at
-) VALUES ('team-a', 'agent-a', 'atm:two', 1, '2026-07-13T00:01:00Z', '2026-07-13T00:01:00Z');
-"#;
-
-    /// Full legacy row projection used to prove the rebuild is lossless.
-    ///
-    /// Every legacy column is TEXT, so one nullable-string vector per row keeps
-    /// the comparison byte-for-byte without a bespoke row struct.
-    type LegacyRow = Vec<Option<String>>;
-
-    fn in_memory_target(label: &str) -> SharedDbTarget {
-        SharedDbTarget::InMemory {
-            uri: format!(
-                "file:atm-storage-rusqlite-{label}-{}?mode=memory&cache=shared",
-                NEXT_IN_MEMORY_DB_ID.fetch_add(1, Ordering::Relaxed)
-            ),
-        }
-    }
-
-    fn legacy_mail_messages_connection(label: &str) -> (SharedDbTarget, Connection) {
-        let target = in_memory_target(label);
-        let connection = open_connection_for_target(&target).expect("open connection");
-        connection
-            .execute_batch(LEGACY_MAIL_SCHEMA_DDL)
-            .expect("create legacy mail schema");
-        (target, connection)
-    }
-
-    fn fresh_schema_connection(label: &str) -> (SharedDbTarget, Connection) {
-        let target = in_memory_target(label);
-        let mut connection = open_connection_for_target(&target).expect("open connection");
-        ensure_schema(&mut connection, &target).expect("fresh schema");
-        (target, connection)
-    }
-
-    fn message_text_notnull(connection: &Connection) -> i64 {
-        connection
-            .query_row(
-                r#"SELECT "notnull" FROM pragma_table_info('mail_messages') WHERE name = 'message_text';"#,
-                [],
-                |row| row.get(0),
-            )
-            .expect("probe message_text nullability")
-    }
-
-    fn foreign_keys_enabled(connection: &Connection) -> i64 {
-        connection
-            .query_row("PRAGMA foreign_keys;", [], |row| row.get(0))
-            .expect("read foreign_keys pragma")
-    }
-
-    fn foreign_key_violations(connection: &Connection) -> i64 {
-        connection
-            .query_row(
-                "SELECT COUNT(1) FROM pragma_foreign_key_check;",
-                [],
-                |row| row.get(0),
-            )
-            .expect("run foreign key check")
-    }
-
-    fn schema_objects(connection: &Connection, object_type: &str) -> Vec<(String, String)> {
-        connection
-            .prepare(
-                "SELECT name, sql FROM sqlite_master
-                 WHERE type = ?1 AND sql IS NOT NULL AND name LIKE '%mail_messages%'
-                 ORDER BY name;",
-            )
-            .expect("prepare sqlite_master query")
-            .query_map([object_type], |row| {
-                Ok((row.get::<_, String>(0)?, row.get::<_, String>(1)?))
-            })
-            .expect("query sqlite_master")
-            .collect::<Result<Vec<_>, _>>()
-            .expect("decode sqlite_master rows")
-    }
-
-    fn decomposed_view_sql(connection: &Connection) -> String {
-        connection
-            .query_row(
-                "SELECT sql FROM sqlite_master WHERE type = 'view' AND name = 'decomposed_messages';",
-                [],
-                |row| row.get(0),
-            )
-            .expect("read decomposed_messages view sql")
-    }
-
-    fn mail_messages_rootpage(connection: &Connection) -> i64 {
-        connection
-            .query_row(
-                "SELECT rootpage FROM sqlite_master WHERE type = 'table' AND name = 'mail_messages';",
-                [],
-                |row| row.get(0),
-            )
-            .expect("read mail_messages rootpage")
-    }
-
-    fn legacy_rows(connection: &Connection) -> Vec<LegacyRow> {
-        connection
-            .prepare(
-                "SELECT team, agent, message_key, envelope_json, from_agent,
-                        source_chat_id, destination_chat_id, message_text, summary,
-                        message_at, message_id, parent_message_id, thread_mode,
-                        recorded_at
-                 FROM mail_messages ORDER BY team, agent, message_key;",
-            )
-            .expect("prepare legacy row query")
-            .query_map([], |row| {
-                (0..14)
-                    .map(|index| row.get(index))
-                    .collect::<Result<_, _>>()
-            })
-            .expect("query legacy rows")
-            .collect::<Result<Vec<_>, _>>()
-            .expect("decode legacy rows")
-    }
-
-    #[test]
-    fn legacy_not_null_message_text_is_rebuilt_without_losing_rows_or_schema_objects() {
-        let (target, mut connection) = legacy_mail_messages_connection("message-text-rebuild");
-        connection
-            .execute_batch(LEGACY_MAIL_ROWS_DDL)
-            .expect("seed legacy rows");
-        assert_eq!(message_text_notnull(&connection), 1);
-        let before = legacy_rows(&connection);
-
-        ensure_schema(&mut connection, &target).expect("migrate legacy database");
-
-        assert_eq!(
-            message_text_notnull(&connection),
-            0,
-            "the rebuild must relax message_text to nullable"
-        );
-        assert_eq!(
-            legacy_rows(&connection),
-            before,
-            "every legacy column value must survive the rebuild unchanged"
-        );
-        let state: (i64, Option<String>) = connection
-            .query_row(
-                "SELECT read, acknowledged_at FROM mail_message_states
-                 WHERE team = 'team-a' AND agent = 'agent-a' AND message_key = 'atm:two';",
-                [],
-                |row| Ok((row.get(0)?, row.get(1)?)),
-            )
-            .expect("dependent state row must survive the foreign-key detour");
-        assert_eq!(state, (1, Some("2026-07-13T00:01:00Z".to_owned())));
-        assert_eq!(foreign_key_violations(&connection), 0);
-        assert_eq!(
-            foreign_keys_enabled(&connection),
-            1,
-            "foreign key enforcement must be restored after the rebuild"
-        );
-
-        let (_fresh_target, fresh) = fresh_schema_connection("message-text-fresh");
-        assert_eq!(
-            schema_objects(&connection, "index"),
-            schema_objects(&fresh, "index"),
-            "a migrated database must be index-identical to a fresh one"
-        );
-        assert_eq!(
-            decomposed_view_sql(&connection),
-            decomposed_view_sql(&fresh),
-            "the catalog view must be recreated exactly as a fresh database defines it"
-        );
-    }
-
-    #[test]
-    fn rebuilt_mail_messages_accepts_the_decomposed_admission_null_update() {
-        let (target, mut connection) = legacy_mail_messages_connection("message-text-null-update");
-        connection
-            .execute_batch(LEGACY_MAIL_ROWS_DDL)
-            .expect("seed legacy rows");
-        ensure_schema(&mut connection, &target).expect("migrate legacy database");
-
-        // Mirrors the decomposed admission write in `writer::ops`, which is the
-        // exact statement that failed on legacy databases.
-        let changed = connection
-            .execute(
-                "UPDATE mail_messages
-                 SET template_sha = 'sha-1', vars_json = '{}', category = 'assignment',
-                     tags_json = '[]', content_format = 'markdown', message_text = NULL
-                 WHERE message_key = 'atm:one' AND template_sha IS NULL",
-                [],
-            )
-            .expect("decomposed admission must be able to null message_text");
-        assert_eq!(changed, 1);
-        let stored: Option<String> = connection
-            .query_row(
-                "SELECT message_text FROM mail_messages WHERE message_key = 'atm:one';",
-                [],
-                |row| row.get(0),
-            )
-            .expect("read migrated row");
-        assert_eq!(stored, None);
-    }
-
-    #[test]
-    fn message_text_rebuild_is_skipped_once_the_column_is_already_nullable() {
-        let (target, mut connection) = legacy_mail_messages_connection("message-text-idempotent");
-        connection
-            .execute_batch(LEGACY_MAIL_ROWS_DDL)
-            .expect("seed legacy rows");
-        ensure_schema(&mut connection, &target).expect("first migration");
-        let migrated_rootpage = mail_messages_rootpage(&connection);
-
-        ensure_schema(&mut connection, &target).expect("second migration");
-        assert_eq!(
-            mail_messages_rootpage(&connection),
-            migrated_rootpage,
-            "a migrated database must not be rebuilt a second time"
-        );
-
-        let (fresh_target, mut fresh) = fresh_schema_connection("message-text-fresh-idempotent");
-        let fresh_rootpage = mail_messages_rootpage(&fresh);
-        ensure_schema(&mut fresh, &fresh_target).expect("re-run fresh schema");
-        assert_eq!(
-            mail_messages_rootpage(&fresh),
-            fresh_rootpage,
-            "a fresh database must never be rebuilt"
-        );
-    }
-
-    #[test]
-    fn message_text_rebuild_preserves_the_message_id_uniqueness_guarantee() {
-        let (target, mut connection) = legacy_mail_messages_connection("message-text-unique");
-        connection
-            .execute_batch(LEGACY_MAIL_ROWS_DDL)
-            .expect("seed legacy rows");
-        let duplicate = "INSERT INTO mail_messages(
-                team, agent, message_key, envelope_json, from_agent, message_text,
-                message_at, message_id
-             ) VALUES ('team-a', 'agent-a', 'atm:duplicate', '{}', 'sender', 'body',
-                       '2026-07-13T00:00:05Z', 'msg-1');";
-        connection
-            .execute_batch(duplicate)
-            .expect_err("legacy unique index must reject a duplicate message_id");
-
-        ensure_schema(&mut connection, &target).expect("migrate legacy database");
-
-        connection
-            .execute_batch(duplicate)
-            .expect_err("the recreated unique index must still reject a duplicate message_id");
-    }
-
-    #[test]
-    fn message_text_rebuild_handles_an_empty_legacy_table() {
-        let (target, mut connection) = legacy_mail_messages_connection("message-text-empty");
-        ensure_schema(&mut connection, &target).expect("migrate empty legacy database");
-
-        assert_eq!(message_text_notnull(&connection), 0);
-        assert_eq!(
-            connection
-                .query_row("SELECT COUNT(1) FROM mail_messages;", [], |row| row
-                    .get::<_, i64>(0))
-                .expect("count migrated rows"),
-            0
-        );
-        assert_eq!(foreign_keys_enabled(&connection), 1);
-    }
-
-    #[test]
-    fn message_text_rebuild_lands_the_fresh_column_list_in_the_fresh_order() {
-        let (target, mut connection) = legacy_mail_messages_connection("message-text-columns");
-        connection
-            .execute_batch(LEGACY_MAIL_ROWS_DDL)
-            .expect("seed legacy rows");
-        assert!(
-            !table_column_names(&connection, &target, "mail_messages")
-                .expect("legacy columns")
-                .contains(&"template_sha".to_owned()),
-            "the legacy fixture must predate the additive template columns"
-        );
-
-        ensure_schema(&mut connection, &target).expect("migrate legacy database");
-
-        let (fresh_target, fresh) = fresh_schema_connection("message-text-columns-fresh");
-        assert_eq!(
-            table_column_names(&connection, &target, "mail_messages").expect("migrated columns"),
-            table_column_names(&fresh, &fresh_target, "mail_messages").expect("fresh columns"),
-            "a migrated table must expose the fresh column list in the fresh order"
-        );
-    }
-
-    #[test]
-    fn a_failed_message_text_rebuild_rolls_back_and_restores_foreign_keys() {
-        let (target, mut connection) = legacy_mail_messages_connection("message-text-rollback");
-        connection
-            .execute_batch(LEGACY_MAIL_ROWS_DDL)
-            .expect("seed legacy rows");
-        let before = legacy_rows(&connection);
-        // Occupying the staging name makes the rebuild's `CREATE TABLE` fail
-        // after the view drop, which is the riskiest point to roll back from.
-        connection
-            .execute_batch("CREATE TABLE mail_messages_nullable_rebuild (probe TEXT);")
-            .expect("occupy the rebuild staging table name");
-
-        let failure = ensure_schema(&mut connection, &target)
-            .expect_err("a blocked rebuild must surface as an error");
-        assert!(
-            failure
-                .message()
-                .contains("failed to create the rebuilt mail_messages table"),
-            "the failing stage must be identifiable: {}",
-            failure.message()
-        );
-
-        assert_eq!(
-            message_text_notnull(&connection),
-            1,
-            "the legacy table must be left untouched so a later startup retries"
-        );
-        assert_eq!(legacy_rows(&connection), before);
-        assert_eq!(foreign_keys_enabled(&connection), 1);
-        assert!(
-            connection
-                .query_row(
-                    "SELECT COUNT(1) FROM sqlite_master WHERE type = 'view' AND name = 'decomposed_messages';",
-                    [],
-                    |row| row.get::<_, i64>(0),
-                )
-                .expect("count views")
-                > 0,
-            "the rollback must restore the dropped catalog view"
-        );
-    }
-
-    #[test]
-    fn a_legacy_table_with_an_unknown_column_is_rejected_instead_of_losing_data() {
-        let (target, mut connection) = legacy_mail_messages_connection("message-text-extra-column");
-        connection
-            .execute_batch("ALTER TABLE mail_messages ADD COLUMN operator_annotation TEXT NULL;")
-            .expect("add an unexpected legacy column");
-        connection
-            .execute_batch(LEGACY_MAIL_ROWS_DDL)
-            .expect("seed legacy rows");
-        let before = legacy_rows(&connection);
-
-        let failure = ensure_schema(&mut connection, &target)
-            .expect_err("an unknown legacy column must block the rebuild");
-        assert!(
-            failure.message().contains("operator_annotation"),
-            "the error must name the column that would be dropped: {}",
-            failure.message()
-        );
-
-        assert_eq!(message_text_notnull(&connection), 1);
-        assert_eq!(legacy_rows(&connection), before);
-        assert_eq!(foreign_keys_enabled(&connection), 1);
-        assert!(
-            table_column_names(&connection, &target, "mail_messages")
-                .expect("columns")
-                .contains(&"operator_annotation".to_owned())
-        );
     }
 }
